+++ src/nl/jchmb/ai/search/cost/CostFunction.java
@@ -6,5 +6,5 @@ import nl.jchmb.ai.search.node.Node;
  * Calculates the cost of a given state.
  */
 public interface CostFunction<T> {
-	public int getCost(Node<T> object);
+	public float getCost(Node<T> object);
 }
