+++ src/nl/jchmb/ai/search/cost/heuristic/DijkstraHeuristic.java
@@ -2,7 +2,7 @@ package nl.jchmb.ai.search.cost.heuristic;
 
 public class DijkstraHeuristic<T> implements Heuristic<T> {
 	@Override
-	public int estimate(T state) {
+	public float estimate(T state) {
 		return 0;
 	}
 }
