+++ src/nl/jchmb/ai/search/cost/heuristic/Heuristic.java
@@ -1,5 +1,5 @@
 package nl.jchmb.ai.search.cost.heuristic;
 
 public interface Heuristic<T> {
-	public int estimate(T state);
+	public float estimate(T state);
 }
